added more functionality to augeochem client
- fixed incorrect type on lat/lon
- added fetch lookups to client in order to fetch the lookup tables that will used to find the correct ids for use when matching field names
- added datapackage ids to appsettings

+++ src/ImuExports/Tasks/AusGeochem/Models/Api/LocationDto.cs
@@ -26,13 +26,13 @@ public class LocationDto
         public int? Id { get; set; }
 
         [JsonPropertyName("lat")]
-        public long? Lat { get; set; }
+        public double? Lat { get; set; }
 
         [JsonPropertyName("latLonPrecision")]
         public int? LatLonPrecision { get; set; }
 
         [JsonPropertyName("lon")]
-        public long? Lon { get; set; }
+        public double? Lon { get; set; }
 
         [JsonPropertyName("name")]
         public string Name { get; set; }
